requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
configobj==5.0.6
discord==1.7.3
discord.py==1.7.3
docopt==0.6.2
gdata==2.0.18
greenlet==1.1.0
idna==3.2
keyring==4.1.1
msgpack==1.0.2
multidict==5.1.0
numpy==1.21.1
Pillow==8.3.1
psutil==5.8.0
pycodestyle==2.7.0
pynvim==0.4.3
python-dateutil==2.8.2
python-dotenv==0.19.0
python-xlib==0.31
pytz==2021.1
requests==2.26.0
six==1.15.0
taskw==0.8.6
taskwarrior-inthe.am==1.1.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
